jmDNS
jmDNS implemented for server and client on all services. Issued ran into with serviceInfo.getPort() on windows, so see comments and report on how it was tested and ran on Linux. Hard-coded port used for windows.

Also added logger into client classes for all services and try-catch implemented, catch any StatusRunTime exception and this will be logged.